Rename example-styx-plugin to styx-test-plugin.
