+++ system-tests/e2e-suite/src/test/java/com/hotels/styx/ExamplePluginJarLocation.java
@@ -1,5 +1,5 @@
 /*
-  Copyright (C) 2013-2018 Expedia Inc.
+  Copyright (C) 2013-2019 Expedia Inc.
 
   Licensed under the Apache License, Version 2.0 (the "License");
   you may not use this file except in compliance with the License.
@@ -45,11 +45,11 @@ public final class ExamplePluginJarLocation {
     }
 
     static Path examplePluginJarLocation() throws IOException {
-        return jarLocation("example-styx-plugin");
+        return jarLocation("styx-test-plugin");
     }
 
     static Path exampleDependencyJarLocation() throws IOException {
-        return jarLocation("example-styx-plugin-dependencies");
+        return jarLocation("styx-test-plugin-dependencies");
     }
 
     // module must be adjacent to e2e-suite
