+++ system-tests/pom.xml
@@ -26,8 +26,8 @@
   <name>Styx - System Tests</name>
 
   <modules>
-    <module>example-styx-plugin-dependencies</module>
-    <module>example-styx-plugin</module>
+    <module>styx-test-plugin-dependencies</module>
+    <module>styx-test-plugin</module>
     <module>e2e-testsupport</module>
     <module>e2e-suite</module>
     <module>example-backend-provider</module>
