+++ system-tests/styx-test-plugin/src/main/java/loadtest/plugins/AsyncRequestContentDecoderPluginFactory.java
@@ -1,5 +1,5 @@
 /*
-  Copyright (C) 2013-2018 Expedia Inc.
+  Copyright (C) 2013-2019 Expedia Inc.
 
   Licensed under the Apache License, Version 2.0 (the "License");
   you may not use this file except in compliance with the License.
@@ -15,42 +15,50 @@
  */
 package loadtest.plugins;
 
-import com.hotels.styx.api.Eventual;
 import com.hotels.styx.api.LiveHttpRequest;
 import com.hotels.styx.api.LiveHttpResponse;
+import com.hotels.styx.api.Eventual;
 import com.hotels.styx.api.plugins.spi.Plugin;
 import com.hotels.styx.api.plugins.spi.PluginFactory;
-import com.hotels.styx.common.CompletableFutures;
 
 import java.util.concurrent.CompletableFuture;
 
+import static com.hotels.styx.common.CompletableFutures.fromSingleObservable;
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
 import static rx.Observable.timer;
 
-public class AsyncResponsePluginFactory implements PluginFactory {
+public class AsyncRequestContentDecoderPluginFactory implements PluginFactory {
+
     @Override
-    public Plugin create(Environment environment) {
+    public Plugin create(PluginFactory.Environment environment) {
         AsyncPluginConfig config = environment.pluginConfig(AsyncPluginConfig.class);
-        return new AsyncResponseInterceptor(config);
+        return new AsyncRequestContentDecoder(config);
     }
 
-    private static class AsyncResponseInterceptor extends AbstractTestPlugin {
+    private static class AsyncRequestContentDecoder extends AbstractTestPlugin {
         private final AsyncPluginConfig config;
 
-        AsyncResponseInterceptor(AsyncPluginConfig config) {
+        AsyncRequestContentDecoder(AsyncPluginConfig config) {
             this.config = config;
         }
 
         @Override
         public Eventual<LiveHttpResponse> intercept(LiveHttpRequest request, Chain chain) {
-            return chain.proceed(request)
-                    .flatMap(response -> Eventual.from(processAsynchronously(response, config.delayMillis())));
-        }
-
-        private static CompletableFuture<LiveHttpResponse> processAsynchronously(LiveHttpResponse response, int delayMillis) {
-            return CompletableFutures.fromSingleObservable(timer(delayMillis, MILLISECONDS)
-                    .map(x -> response));
+            return request.aggregate(config.maxContentLength())
+                            .flatMap(fullHttpRequest -> Eventual.from(asyncOperation(config.delayMillis())))
+                            .map(outcome -> request.newBuilder().header("X-Outcome", outcome.result()))
+                            .flatMap(x -> chain.proceed(request));
         }
     }
-}
 
+    private static CompletableFuture<Outcome> asyncOperation(long delay) {
+        return fromSingleObservable(timer(delay, MILLISECONDS)).thenApply(x -> new Outcome());
+    }
+
+    private static class Outcome {
+        int result() {
+            return 1;
+        }
+    }
+
+}
