+++ system-tests/styx-test-plugin/src/main/java/testgrp/TestPluginConfig.java
